# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API documentation",
        "contact": {
            "email": "benabdeljelilaicha@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/events": {
            "get": {
                "tags": [
                    "Code challenge (Events)"
                ],
                "summary": "Get all events data",
                "description": "Returns all events data",
                "operationId": "getAllEvents",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Code challenge (Events)"
                ],
                "summary": "Set events data",
                "description": "Set events data into database",
                "operationId": "setEvents",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/search/events": {
            "get": {
                "tags": [
                    "Code challenge (Events)"
                ],
                "summary": "Get filtred events",
                "description": "Returns filtred events data",
                "operationId": "searchEvents",
                "parameters": [
                    {
                        "name": "employeeName",
                        "in": "query",
                        "description": "Employee name"
                    },
                    {
                        "name": "eventName",
                        "in": "query",
                        "description": "Event name"
                    },
                    {
                        "name": "eventDate",
                        "in": "query",
                        "description": "Event date",
                        "example": "2019-09-04"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Code challenge (Events)",
            "description": "API Endpoints of Projects"
        }
    ],
    "security": [
        []
    ]
}